feat(update-designables): changed all controls with ibproperties, separated code for better readability.
