--- CustomAVPlayer/Enums/IconSize.swift
@@ -1,14 +1,0 @@
-//
-//  IconSize.swift
-//  CustomAVPlayer
-//
-//  Created by Himanshu on 7/26/23.
-//
-
-import Foundation
-
-public enum IconSize: CGFloat {
-    case small = 12
-    case medium = 24
-    case large = 36
-}
