--- CustomAVPlayer/Enums/SliderPosition.swift
@@ -1,14 +1,0 @@
-//
-//  SliderPosition.swift
-//  CustomAVPlayer
-//
-//  Created by Himanshu on 7/26/23.
-//
-
-import Foundation
-
-public enum SliderPosition {
-    case defaultPosition
-    case bottomSticked
-    case hidden
-}
