+++ CustomAVPlayer/Static/Constants.swift
@@ -8,14 +8,7 @@
 import Foundation
 
 struct ConstantString {
-    
-    static let videoPlayerView = "VideoPlayerView"
-    
-    static let fatalErrorLoadingNib = "init(coder:) has not been implemented"
-    
     static let numberFormat00 = "%02d"
-    
     static let transformRotation = "transform.rotation"
-    
     static let timeControlStatus = "timeControlStatus"
 }
